refactor: adapt common folder structure and naming conventions
